requirements.txt
streamlit 
python-docx 
PyMuPDF 
pytesseract 
Pillow
python-dotenv

openai 
groq